+++ system/src/Grav/Common/Twig/Node/TwigNodeThrow.php
@@ -1,0 +1,40 @@
+<?php
+/**
+ * @package    Grav.Common.Twig
+ *
+ * @copyright  Copyright (C) 2015 - 2018 Trilby Media, LLC. All rights reserved.
+ * @license    MIT License; see LICENSE file for details.
+ */
+
+namespace Grav\Common\Twig\Node;
+
+class TwigNodeThrow extends \Twig_Node
+{
+    public function __construct(
+        $code,
+        \Twig_Node $message,
+        $lineno = 0,
+        $tag = null
+    )
+    {
+        parent::__construct(['message' => $message], ['code' => $code], $lineno, $tag);
+    }
+
+    /**
+     * Compiles the node to PHP.
+     *
+     * @param \Twig_Compiler $compiler A Twig_Compiler instance
+     * @throws \LogicException
+     */
+    public function compile(\Twig_Compiler $compiler)
+    {
+        $compiler->addDebugInfo($this);
+
+        $compiler
+            ->write('throw new \RuntimeException(')
+            ->subcompile($this->getNode('message'))
+            ->write(', ')
+            ->write($this->getAttribute('code') ?: 500)
+            ->write(");\n");
+    }
+}